testing

+++ src/base/TensorFFT256.cu
@@ -237,8 +237,8 @@ __global__ void TensorFFT256(__half* input_data_RE, __half* input_data_IM,
     //            static_cast<__half>(128.0));
     float phase = (static_cast<float>(inter_warp_id_16 * j) * M_PI) / 128.0;
 
-    __half twiddle_RE = cosf(phase);
-    __half twiddle_IM = -sinf(phase);
+    __half twiddle_RE = cos(phase);
+    __half twiddle_IM = -sin(phase);
 
     __half input_RE = buffer_RE[buffer_array_id];
     __half input_IM = buffer_IM[buffer_array_id];
